+++ CHANGELOG.md
@@ -9,11 +9,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Added
 - federated voter count of polls is now parsed and federated out too;
     this fixes vote percetanges for new and refreshed remote multi-selection polls
+- new config options to restrict unauthenticated search API access under `:pleroma, :restrict_unauthenticated, :search`
 
 ### Fixed
+- fixed status search not respecting `resolve=false`
 
 ### Changed
 - New installations (not existing instances) now default to the `simple` full-text-search config
+- Unauthenticated search requests now by default force-disable remote fetches and pagination
 
 
 ## 2026.05 (3.19.0)